+++ components/ads-client/src/ffi.rs
@@ -141,7 +141,7 @@ impl MozAdsClientBuilder {
                 .clone()
                 .map(MozAdsContextIdProviderWrapper::new)
                 .map(Into::into),
-            environment: inner.environment.unwrap_or_default().into(),
+            environment: inner.environment.clone().unwrap_or_default().into(),
             telemetry: telemetry.clone(),
             store_config: inner.store_config.clone().map(Into::into),
         };
@@ -189,13 +189,14 @@ impl MozAdsClientBuilder {
     }
 }
 
-#[derive(Clone, Copy, Debug, Default, uniffi::Enum, Eq, PartialEq)]
+#[derive(Clone, Debug, Default, uniffi::Enum, Eq, PartialEq)]
 pub enum MozAdsEnvironment {
     #[default]
     Prod,
     Staging,
     #[cfg(test)]
     Test,
+    Custom(AdsClientUrl),
 }
 
 #[derive(Clone, uniffi::Record)]
@@ -398,6 +399,7 @@ impl From<Environment> for MozAdsEnvironment {
             Environment::Staging => MozAdsEnvironment::Staging,
             #[cfg(test)]
             Environment::Test => MozAdsEnvironment::Test,
+            Environment::Custom(url) => MozAdsEnvironment::Custom(url),
         }
     }
 }
@@ -409,6 +411,7 @@ impl From<MozAdsEnvironment> for Environment {
             MozAdsEnvironment::Staging => Environment::Staging,
             #[cfg(test)]
             MozAdsEnvironment::Test => Environment::Test,
+            MozAdsEnvironment::Custom(url) => Environment::Custom(url),
         }
     }
 }