+++ components/ads-client/src/mars/environment.rs
@@ -11,13 +11,14 @@ static MARS_API_ENDPOINT_PROD: Lazy<Url> = Lazy::new(|| url!("https://ads.mozill
 
 static MARS_API_ENDPOINT_STAGING: Lazy<Url> = Lazy::new(|| url!("https://ads.allizom.org/v1/"));
 
-#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+#[derive(Clone, Debug, Default, Eq, Hash, PartialEq)]
 pub enum Environment {
     #[default]
     Prod,
     Staging,
     #[cfg(test)]
     Test,
+    Custom(Url),
 }
 
 impl Environment {
@@ -36,6 +37,7 @@ impl Environment {
             Environment::Staging => MARS_API_ENDPOINT_STAGING.clone(),
             #[cfg(test)]
             Environment::Test => Url::parse(&mockito::server_url()).unwrap(),
+            Environment::Custom(url) => url,
         }
     }
 }