+++ components/support/sql/src/lib.rs
@@ -12,9 +12,10 @@ pub mod debug_tools;
 
 mod each_chunk;
 mod lazy;
-mod maintenance;
+pub mod maintenance;
 mod maybe_cached;
 pub mod open_database;
+pub mod path;
 mod repeat;
 
 pub use conn_ext::*;